+++ Package.swift
@@ -15,8 +15,9 @@ let package = Package(
     .package(name: "MungoHealer", url: "https://github.com/Flinesoft/MungoHealer.git", from: "0.3.4"),
     .package(name: "Rainbow", url: "https://github.com/onevcat/Rainbow.git", from: "3.1.5"),
     .package(name: "SwiftCLI", url: "https://github.com/jakeheis/SwiftCLI.git", from: "6.0.3"),
-    .package(name: "Toml", url: "https://github.com/jdfergason/swift-toml.git", .branch("master")),
     .package(name: "SwiftSyntax", url: "https://github.com/apple/swift-syntax.git", from: "0.50500.0"),
+    .package(name: "SwiftyXML", url: "https://github.com/chenyunguiMilook/SwiftyXML.git", from: "3.1.0"),
+    .package(name: "Toml", url: "https://github.com/jdfergason/swift-toml.git", .branch("master")),
   ],
   targets: [
     .executableTarget(
@@ -32,6 +33,7 @@ let package = Package(
         "Rainbow",
         "SwiftCLI",
         "SwiftSyntax",
+        "SwiftyXML",
         "Toml",
       ]
     ),